pyproject.toml
[project]
name = "chatbot-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "anthropic>=0.46.0",
    "gradio==5.17.1",
    "python-dotenv>=1.0.1",
]
